feat: stellar-config crate
Move all config logic into its own crate
