+++ cmd/crates/soroban-spec-tools/Cargo.toml
@@ -29,11 +29,3 @@ hex = { workspace = true }
 wasmparser = { workspace = true }
 base64 = { workspace = true }
 thiserror = "1.0.31"
-# soroban-ledger-snapshot = { workspace = true }
-# soroban-sdk = { workspace = true }
-# sep5 = { workspace = true }
-
-
-[dev-dependencies]
-which = { workspace = true }
-tokio = "1.28.1"
