+++ cmd/crates/soroban-test/Cargo.toml
@@ -26,6 +26,7 @@ soroban-sdk = { workspace = true }
 sep5 = { workspace = true }
 soroban-cli = { workspace = true }
 soroban-rpc = { workspace = true }
+config = { workspace = true }
 
 thiserror = "1.0.31"
 sha2 = "0.10.6"
