+++ cmd/crates/stellar-config/Cargo.toml
@@ -1,0 +1,74 @@
+[package]
+name = "stellar-config"
+description = "Tools for managing Stellar configuration"
+homepage = "https://github.com/stellar/soroban-cli"
+repository = "https://github.com/stellar/soroban-cli"
+authors = ["Stellar Development Foundation <info@stellar.org>"]
+readme = "README.md"
+license = "Apache-2.0"
+version.workspace = true
+edition = "2021"
+rust-version.workspace = true
+
+[dependencies]
+stellar-xdr = { workspace = true, features = ["cli"] }
+soroban-env-host = { workspace = true }
+soroban-spec = { workspace = true }
+stellar-strkey = { workspace = true }
+soroban-sdk = { workspace = true }
+soroban-rpc = { workspace = true }
+# config_inner =  {version = "0.14.0", package = "config", features = ["toml"]}
+
+clap = { workspace = true, features = [
+  "derive",
+  "env",
+  "deprecated",
+  "string",
+] }
+
+base64 = { workspace = true }
+thiserror = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+serde-aux = { workspace = true }
+hex = { workspace = true }
+rpassword = "7.2.0"
+tokio = { version = "1", features = ["full"] }
+termcolor = { workspace = true }
+termcolor_output = { workspace = true }
+rand = "0.8.5"
+
+sha2 = { workspace = true }
+
+ed25519-dalek = "=2.0.0"
+jsonrpsee-http-client = "0.20.1"
+jsonrpsee-core = "0.20.1"
+hyper = "0.14.27"
+hyper-tls = "0.5"
+http = "0.2.9"
+
+chrono = "0.4.27"
+
+dirs = "4.0.0"
+toml = "0.5.9"
+itertools = "0.10.5"
+
+sep5 = { workspace = true }
+ethnum = { workspace = true }
+clap-markdown = { version = "0.1.3", optional = true }
+
+
+
+tracing = { workspace = true }
+
+tempfile = "3.8.1"
+futures-util = "0.3.30"
+
+
+[build-dependencies]
+crate-git-revision = "0.0.4"
+serde.workspace = true
+thiserror.workspace = true
+
+
+
