+++ cmd/crates/stellar-config/README.md
@@ -1,0 +1,3 @@
+# stellar-config
+
+Configuration for stellar, includes, keys, networks and contract aliases.
