+++ cmd/soroban-cli/src/commands/contract/restore.rs
@@ -1,4 +1,4 @@
-use std::{fmt::Debug, path::Path, str::FromStr};
+use std::{fmt::Debug, str::FromStr};
 
 use clap::{command, Parser};
 use soroban_env_host::xdr::{
@@ -9,15 +9,13 @@ use soroban_env_host::xdr::{
 };
 use stellar_strkey::DecodeError;
 
+use config::locator;
+
 use crate::{
-    commands::{
-        config::{self, locator},
-        contract::extend,
-        global, NetworkRunnable,
-    },
+    commands::{contract::extend, global, NetworkRunnable},
     key,
     rpc::{self, Client},
-    wasm, Pwd,
+    wasm,
 };
 
 #[derive(Parser, Debug, Clone)]
@@ -43,12 +41,6 @@ impl FromStr for Cmd {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         use clap::{CommandFactory, FromArgMatches};
         Self::from_arg_matches_mut(&mut Self::command().get_matches_from(s.split_whitespace()))
-    }
-}
-
-impl Pwd for Cmd {
-    fn set_pwd(&mut self, pwd: &Path) {
-        self.config.set_pwd(pwd);
     }
 }
