+++ cmd/soroban-cli/src/commands/events.rs
@@ -3,10 +3,9 @@ use std::io;
 
 use soroban_env_host::xdr::{self, Limits, ReadXdr};
 
-use super::{
-    config::{self, locator},
-    global, network, NetworkRunnable,
-};
+use config::{locator, network};
+
+use super::{global, NetworkRunnable};
 use crate::{rpc, utils};
 
 #[derive(Parser, Debug, Clone)]
