+++ cmd/soroban-cli/src/commands/keys/fund.rs
@@ -1,6 +1,6 @@
 use clap::command;
 
-use crate::commands::network;
+use config::network;
 
 use super::address;
